updated login page with shoelace components and converted to folder
