+++ client/src/pages/Login/Login.module.css
@@ -1,0 +1,9 @@
+.signupCTA {
+  text-align: center;
+  margin-top: 2rem;
+}
+
+.formContainer {
+  max-width: 550px;
+  margin: auto;
+}
